搞定了文章发布功能，明天修改文章展示，以及增加本地图片上传和文章详情页的展示

+++ dao/src/main/resources/tb_article_tag_middle.sql
@@ -2,7 +2,7 @@
 CREATE TABLE tb_article_tag_middle(
 article_tag_middle_id INT(11) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '文章标签中间表主键',
 article_id BIGINT(20) UNSIGNED NOT NULL COMMENT '文章id',
-tag_id INT(11) UNSIGNED NOT NULL COMMENT '标签id',
+tag_id BIGINT(20) UNSIGNED NOT NULL COMMENT '标签id',
 PRIMARY KEY (article_tag_middle_id)
 )ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='文章标签中间表';
